Complexity Improvement
Identified an early exit while searching for the smallest prime factor of a number. As the smallest prime factor will be less than the square root of the non prime number itself we just need to search until the pre calculated primes start getting bigger than sqrt(n). After that we are sure that the number must be a prime itself.